backend/pyproject.toml
[project]
name = "budgeter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "django>=5.2.3",
    "django-allauth>=65.14.0",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
